Fixed multi project FCERM1 export (PM-297) and downloads issue
