+++ app/views/pafs_core/projects/index.html.erb
@@ -14,7 +14,7 @@
 
   <% if ea_user? %>
     <div class="download-all-proposals-link">
-      <%= link_to t('download_all_proposals_label'), pafs_core.all_downloads_path %>
+      <%= link_to t('download_all_proposals_label'), pafs_core.multi_downloads_path %>
     </div>
   <% else %>
     <div class="create-new-project">
